requirements.txt
streamlit
yt-dlp
opencv-python
Pillow
spacy
torch
transformers
langdetect
tqdm